[kernel/time] Fix Bluetooth audio stuttering
Note: This taken from: @https://github.com/GrazieMaldini/redkancut/commit/2683821fe2a22b368a6de04acd9a4356508558c3

Taken from http://tinyurl.com/jtb22tp. See that thread for more details.
Original author did not have a Git commit for this, so no cherry pick.
get_monotonic_boottime() is used when playing audio with Bluetooth, and
apparently short repeated calls to it can cause strange values. This
is caused by an overflow of the long value used to store time, since the
values can get really big.

+++ kernel/time/timekeeping.c
@@ -1693,7 +1693,7 @@ struct timespec get_monotonic_coarse(void)
 	} while (read_seqcount_retry(&tk_core.seq, seq));
 
 	set_normalized_timespec64(&now, now.tv_sec + mono.tv_sec,
-				now.tv_nsec + mono.tv_nsec);
+				(s64)now.tv_nsec + (s64)mono.tv_nsec);
 
 	return timespec64_to_timespec(now);
 }
